filter before dialing

+++ config/config.go
@@ -635,8 +635,8 @@ type P2PConfig struct { //nolint: maligned
 	SameRegion                           bool    `mapstructure:"same_region"`
 	MyRegion                             string  `mapstructure:"my_region"`
 	MaxPercentPeersInSameRegion          float64 `mapstructure:"max_percent_peers_in_same_region"`
-	CurrentNumOutboundPeersInOtherRegion float64 `mapstructure:"current_num_outbound_peers_in_other_region"`
-	CurrentNumInboundPeersInOtherRegion  float64 `mapstructure:"current_num_inbound_peers_in_other_region"`
+	CurrentNumOutboundPeersInOtherRegion int     `mapstructure:"current_num_outbound_peers_in_other_region"`
+	CurrentNumInboundPeersInOtherRegion  int     `mapstructure:"current_num_inbound_peers_in_other_region"`
 }
 
 // DefaultP2PConfig returns a default configuration for the peer-to-peer layer
